requirements.txt
google-generativeai
langchain
langchain-core
langgraph
python-dotenv
jira
pydantic
fastapi
uvicorn
python-dateutil
markdownify
html2text
msgraph-sdk
azure-identity
schedule
langchain-google-genai
langchain_community
atlassian-python-api
google-genai
langfuse
slack_sdk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain-google-community[gmail]